% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_results.R
\name{compute_fitted_sim}
\alias{compute_fitted_sim}
\title{Compute the True Values to Compare with Fitted Values}
\usage{
compute_fitted_sim(fitted_cu, I = 10, J = 16, reps = 5, nested = FALSE)
}
\arguments{
\item{fitted_cu}{Object saved from the simulation.}

\item{I}{Number of levels for component B. Defaults to 9.}

\item{J}{Number of levels for component C. Defaults to 16.}

\item{reps}{Number of repetitions of the levels of B and C. Defaults to 5.}

\item{nested}{TRUE if the model component C is nested. Defaults to FALSE.}
}
\description{
This is an internal function. It takes the fitted curves object of the
simulation and recalculates the true curves.
}
