% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_mfpcs.R
\name{compute_var}
\alias{compute_var}
\title{Compute the Number of FPCs needed}
\usage{
compute_var(sigma_sq, values, norms_sq, mfpc_cutoff)
}
\arguments{
\item{sigma_sq}{Vector containing the estimated variances on each dimension.}

\item{values}{List containing the multivariate Eigenvalues for each variance
component.}

\item{norms_sq}{Vector containing the squared norms to be used as weights on
the Eigenvalues.}

\item{mfpc_cutoff}{Pre-specified level of explained variance of results of
MFPCA.}
}
\description{
This is an internal function. The function takes all the information needed
to calculate how many FPCs are needed to reach the pre-specified cutoff
level.
}
