% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfpca.R
\name{conduct_mfpca}
\alias{conduct_mfpca}
\title{Conduct the MFPCA}
\usage{
conduct_mfpca(mfpca_info, mfpc_weight)
}
\arguments{
\item{mfpca_info}{Object containing all the neccessary information for the
MFPCA. List as given by the output of prepare_mfpca().}
}
\description{
This is an internal function contained in the multiFAMM function. This step
uses the information from the univariate FLMMs for the MFPCA. It also allows
a single weighting scheme of the MFPCA.
}
\details{
Currently, it is possible to conduct a non-weighted MFPCA (default) as well
as a MFPCA that uses the estimated univariate error variances as weights.
}
