% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_plots.R
\name{covariate_comp_plot_helper}
\alias{covariate_comp_plot_helper}
\title{Create Data to Plot Covariate Effect Comparison Uni-Mul}
\usage{
covariate_comp_plot_helper(dat_m, aco_pr, epg_pr, mul_level, uni_effect)
}
\arguments{
\item{dat_m}{Helper data from multivariate model.}

\item{aco_pr}{Predictions from first univariate model.}

\item{epg_pr}{Predictions from second univariate model. Can be NULL.}

\item{mul_level}{Effect level of the multivariate model to be plotted.}

\item{uni_effect}{Number of the univariate effect to be plotted.}
}
\description{
This is an internal function. It takes a data set of estimated covariate
effects of a multivariate model and compares it to univariately estimated
covariate effects.
}
