% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_plots.R
\name{covariate_plot_helper_2d_transform}
\alias{covariate_plot_helper_2d_transform}
\title{Transform Covaraiate Effect Plot Data for 2D Case}
\usage{
covariate_plot_helper_2d_transform(data, covs = 0L,
  indicator = c("\\\\.x", "\\\\.y"), m_fac = 1.96)
}
\arguments{
\item{data}{Data containing the covariate effect information as provided by
covariate_plot_helper().}

\item{covs}{Covariate effects to be plotted. Can also be a vector if e.g. an
interaction is to be plotted. Then, the interaction should be the last
element of the vector. Defaults to the intercept.}

\item{indicator}{Two element vector containing the suffix indicating the 2D
axis (first for x, then for y axis). Defaults to labels ending with ".x"
and ".y".}

\item{m_fac}{Multiplication factor to represent the confidence interval of
the estimates. Defaults to 1.96.}
}
\description{
This is an internal function. It gives as an output the data set necessary to
plot 2D trajectories of the effect plots.
}
\details{
If a vector is supplied to covs, then the sum of the first elements up to the
second to last element is the given in the variable ... and the whole sum is
given in the variable ...
}
