% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_plots.R
\name{coverage_area_helper}
\alias{coverage_area_helper}
\title{Coverage Area Plot Helper Function}
\usage{
coverage_area_helper(sim_curves, effect_index, ylim = c(-0.5, 0.5),
  by = 1/99, m_fac = 1.96)
}
\arguments{
\item{sim_curves}{The large list of simulation results. Use object$mul.}

\item{ylim}{Two element numeric vector specifying the area for which the
coverage is to be plotted. Defaults to c(-0.5, 0.5).}

\item{by}{Step size for the grid of y values to check for coverage.
Defaults to 1/99.}

\item{m_fac}{Multiplication factor used to create the upper and lower
credibility bounds. Defaults to 1.96 (ca. 95\%).}
}
\description{
This function is an internal function. The function takes one single effect
function from the simulated effects and gives a data.frame which can be used
to plot the coverage area of all simulation runs.
}
