% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_simulation.R
\name{coverage_plot_helper}
\alias{coverage_plot_helper}
\title{Coverage plot helper function}
\usage{
coverage_plot_helper(cov_list, effect_index, dimlabels = c("ACO", "EPG"))
}
\arguments{
\item{cov_list}{List of arrays containing the evaluations if the estimated
coefficient effect of that simulation run is in the credibility bounds as
given by the function create_coverage_array().}

\item{effect_index}{Numeric vector of the index positions of the effects to
be plotted.}

\item{dimlabels}{String vector of labels used in the data set. Defaults to
labels "ACO" and "EPG".}
}
\description{
This is an internal function. The function takes a list of arrays created by
the function create_coverage_array and returns a data.frame ready for
plotting.
}
