% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_simulation.R
\name{create_coverage_array}
\alias{create_coverage_array}
\title{Create the Coverage Array of the Simulation}
\usage{
create_coverage_array(sim_curves, gen_curves, effect_index, m_fac = 1.96)
}
\arguments{
\item{sim_curves}{The large list of simulation results. Use object$mul.}

\item{gen_curves}{The original data generating curve as part of the output of
multifamm:::extract_components(), so use output$cov_preds.}

\item{m_fac}{Multiplication factor used to create the upper and lower
credibility bounds. Defaults to 1.96 (ca. 95\%).}

\item{effec_index}{The index position of the effect to be evaluated in the
gen_curves and sim_curves effect lists. If the intercept is to be
evaluated, this can be specified as 1 or 2 (both scalar and functional
intercept are sumed up).}
}
\description{
This is an internal function. The function takes the index of the covariate
to be evaluated and then checks whether the estimated covariate effect of the
simulation run covers the true data generating effect function. The output is
a logical array where the first dimension gives the dimension of the data,
the second dimension gives the time point to be evaluated and the third
dimension gives the simulation run.
}
