% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_results.R
\name{extract_Covfct_sim}
\alias{extract_Covfct_sim}
\title{Extract the Estimated Covariate Effect for One Specific Model Term}
\usage{
extract_Covfct_sim(term, m_true_comp, cov_preds, se = FALSE)
}
\arguments{
\item{term}{Name of the covariate effect function to extract.}

\item{m_true_comp}{True model component from which to extract the true
observation.}

\item{cov_preds}{List of simulation results from which to extract (as given
by e.g. load_sim_results()).}

\item{se}{TRUE if the standard error is to be extracted instead of the
estimated function. Defaults to FALSE.}
}
\description{
This is an internal function. The function extracts one single effect
function from the simulated effects and adds the true observation as the last
observation.
}
