% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfpca.R
\name{extract_var_info}
\alias{extract_var_info}
\title{Extract Variance Information from MFPCA Object}
\usage{
extract_var_info(MFPC = MFPC)
}
\arguments{
\item{MFPCA}{MFPCA object from which to extract multivariate eigenvalues and
univariate norms.}
}
\description{
This is an internal function contained in the multiFAMM function. This step
allows to extract the information of the total variation in the data (multi-
and univariate).
}
