% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_plots.R
\name{fpc_plot_helper}
\alias{fpc_plot_helper}
\title{Extract the FPCs and Output a Data.Frame}
\usage{
fpc_plot_helper(
  model,
  mcomp,
  component,
  dimlabels,
  two_d = FALSE,
  multi = TRUE,
  m_fac = 2
)
}
\arguments{
\item{model}{The multifamm model from which to extract the estimates.}

\item{mcomp}{The extracted model components as given by the internal function
extract_components().}

\item{component}{Which FPC component to extract. Possible values ("B", "C",
"E")}

\item{dimlabels}{The dimensions of the model.}

\item{two_d}{Is a two dimensional representation (e.g. x and y coordinate) a
better representation for the data. Defaults to FALSE (one variable).}

\item{multi}{Is the model a multifamm or a sparseFLMM model. Defaults to TRUE
(multifamm model).}

\item{m_fac}{Multiplication factor to represent the difference from the
overall mean.}
}
\description{
This is an internal function. It gives as an output the data set necessary to
plot the eigenfunctions. It is meant to be general enough to also include
more than two dimensions.
}
