% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_results.R
\name{funData2DataFrame}
\alias{funData2DataFrame}
\title{Transform a funData object to an Data.Frame}
\usage{
funData2DataFrame(fundata)
}
\arguments{
\item{fundata}{FunData or MultiFundata object to transform.}
}
\value{
A data.frame with four variables
  \itemize{
    \item \code{t} (num): the functional argument.
    \item \code{y} (num): the functional value.
    \item \code{dim} (int): the number of the dimension (position in the
    multiFunData object).
    \item \code{obs} (int): the number of the observation in the funData
    object.}
}
\description{
This is an internal function. It transform a funData object to a data.frame
so that it is easy to plot it using ggplot. It internally checks if the
object is a multiFunData object.
}
