% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_simulation.R
\name{gendata}
\alias{gendata}
\title{Generate Data for Simulation}
\usage{
gendata(I = 10, J = 10, nested = FALSE, num_dim = 2, lamB, lamC, lamE,
  normal = TRUE, sigmasq = list(0.05, 4), dim_indep = TRUE,
  simple_sig = TRUE, mu = function(t) {     t + sin(t) }, N_B = 1,
  N_C = 1, N_E = 1, phiB_fun, phiC_fun, phiE_fun, minsize = 10,
  maxsize = 20, min_grid = 0, max_grid = 1, grid_eval = 100,
  min_visit = 3, max_visit = 3, use_RI = FALSE, covariate = TRUE,
  num_cov = 4, interaction = FALSE, which_inter = matrix(NA),
  model = NULL, center_scores = FALSE, decor_scores = FALSE,
  nested_center = FALSE)
}
\arguments{
\item{I}{Number of levels for first grouping variable (individuals).}

\item{J}{Number of levels for second grouping variable (set to NA for
random intercept design).}

\item{nested}{TRUE: second random effect is nested. FALSE: if it is
a crossed effect. Defaults to FALSE.}

\item{num_dim}{Number of dimensions.}

\item{lamB}{Eigenvalues for first grouping variable.}

\item{lamC}{Eigenvalues for second grouping variable. Can be NULL.}

\item{lamE}{Eigenvalues for curves-specific deviations.}

\item{normal}{TRUE: FPC weights are drawn from a normal distribution.
FALSE: they are drawn from a mixture of normals. Defaults to TRUE.}

\item{sigmasq}{Error variances for each dimension supplied as a list.}

\item{dim_indep}{TRUE: Use rmvnorm to ensure that error variances are
independent across the dimensions. FALSE: Prolonge Jona's implementation
using . Defaults to TRUE.}

\item{simple_sig}{TRUE: Random normal numbers weighted with sigma. FALSE:
mvnorm. Defaults to TRUE.}

\item{mu}{Mean function. Can be NULL if the GAM model is supplied, then
defaults to t + sin(t). If GAM model is supplied, then this argument is
redundant.}

\item{N_B}{Number of FPCs for first grouping variable. Defaults to 1.}

\item{N_C}{Number of FPCs for second grouping variable. Can be 0 and defaults
to 1.}

\item{N_E}{Number of FPCs for curves-specific deviations. Defaults to 1.}

\item{phiB_fun}{Eigenfunctions of the first grouping variable as a
multiFunData object.}

\item{phiC_fun}{Eigenfunctions of the second grouping variable as a
multiFunData object. Can be NULL.}

\item{phiE_fun}{Eigenfunctions of the curve-specific deviations as a
multiFunData object.}

\item{minsize}{Minimal number of scalar observations per curve. Defaults to
10.}

\item{maxsize}{Maximal number of observations per curve. Defaults to 20.}

\item{min_grid}{Minimal value of grid range. Defaults to 0.}

\item{max_grid}{Maximal value of grid range. Defaults to 1.}

\item{grid_eval}{Length of the final grid to evaluate the functions. Defaults
to 100.}

\item{min_visit}{Minimal number of repetitions of each subject-word/session
(first- and second grouping variable) combination (in case of random
intercept design: minimal number of repetitions for first grouping
variable). Defaults to 3.}

\item{max_visit}{Maximal number of repetitions of each subject-word/session
(first- and second grouping variable) combination (in case of random
intercept design: maximal number of repetitions for first grouping
variable). Defaults to 3.}

\item{use_RI}{TRUE: data with a random intercept structure are generated.
FALSE: data with crossed/nested random intercepts structure are generated
(Additional layer "words"/"sessions" included). Defaults to FALSE.}

\item{covariate}{If covariates shall be generated. Defaults to TRUE.}

\item{num_cov}{Number of covariates if covariate = TRUE. Defaults to 4.}

\item{interaction}{TRUE if there are interactions between covariates (as in
the phonetic sparseFLMM). Defaults to FALSE.}

\item{which_inter}{Symmetric matrix specifying the interaction terms (as in
sparseFLMM). Defaults to missing matrix.}

\item{model}{GAM model from which to extract the covariate functions. Can be
NULL if mu is specified.}

\item{center_scores}{TRUE: FPC weights are centered. Defaults to FALSE.}

\item{decor_scores}{TRUE: FPC weights are decorrelated. Defaults to FALSE.}

\item{nested_center}{TRUE: FPC weights of the nested effect are centered
around zero for each subject. Defaults to FALSE.}
}
\description{
This is an internal function. It creates a new data set based on already
given functions.
}
