% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfrm.R
\name{mvfr}
\alias{mvfr}
\title{Multivariate Functional Additive Mixed Model Regression}
\usage{
mvfr(data, fRI_B = FALSE, fRI_C = FALSE, bs = "ps", bf_mean = 8,
  bf_covariates = 8, m_mean = c(2, 3), covariate = FALSE,
  num_covariates = NULL, covariate_form = NULL, interaction = FALSE,
  which_interaction = matrix(NA), bf_covs, m_covs, var_level = 0.99,
  use_famm = FALSE, save_model_famm = FALSE, one_dim = NULL,
  mfpc_cutoff = 0.95, number_mfpc = NULL,
  mfpc_cut_method = c("total_disp", "unidim"),
  final_method = c("w_bam", "bam", "gamm", "gaulss"), ...)
}
\arguments{
\item{data}{Data.table that contains the information with some fixed variable
names, see Details.}

\item{fRI_B}{Boolean for including functional random intercept for individual
(B in Cederbaum). Defaults to \code{FALSE}.}

\item{fRI_C}{Boolean for including functional random intercept
for word (C in Cederbaum). Defaults to \code{FALSE}.}
}
\value{
A list with four elements
  \itemize{
    \item the final multivariate FAMM
    \item the sparseFLMM output for each of the dimensions
    \item the MFPC output
    \item the data used to fit the model}
}
\description{
This is the main function of the package and fits the multivariate functional
additive regression model with potentially nested or crossed functional
random intercepts.
}
\details{
Expand the method proposed by Fabian Scheipl to incorporate the variance
decomposition developed by Cederbaum et al. (2016). To account for the
correlation between the dimensions, the MFPCA approach by Happ and Greven
(2016) is applied.

The data set has to be of the following format:
\itemize{
  \item y_vec (numeric): vector of response values
  \item t (numeric): observation point locations
  \item n_long (integer): curve identification
  \item subject_long (integer): subject identification (NEEDS TO BE
    SPECIFIED)
  \item word_long (integer): word identification
  \item combi_long (integer): repetition
  \item dim (factor): level of the dimension
  \item covariate.1 (numeric): potential covariate(s) named with trailing
    1,2,3,...
}
}
