% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_results.R
\name{my_corr}
\alias{my_corr}
\title{Correlation Plots for Univariate and Multivariate Scores}
\usage{
my_corr(data, method = c("pairwise", "pearson"), cor_matrix = NULL,
  nbreaks = NULL, digits = 2, name = "", low = "#3B9AB2",
  mid = "#EEEEEE", high = "#F21A00", midpoint = 0, palette = NULL,
  geom = "tile", min_size = 2, max_size = 6, label = FALSE,
  label_alpha = FALSE, label_color = "black", label_round = 1,
  label_size = 3, limits = c(-1, 1), drop = is.null(limits) ||
  identical(limits, FALSE), layout.exp = 0, legend.position = "right",
  legend.size = 6, text_size = 2.7, xmin = 3.5, xmax = 7.5,
  ymin = 0.5, ymax = 3.5, xmin_d = 5.5, xmax_d = 7.5, ymin_d = 3.5,
  ymax_d = 5.5, ...)
}
\description{
This is an internal function. This code is an adjustment from the package
GGally version 1.3.2 of the function ggcorr.
}
