% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phonetic_subset}
\alias{phonetic_subset}
\title{Phonetic data (subset)}
\format{
A data.frame with 1336 observations and 12 variables.
}
\source{
Pouplier, Marianne and Hoole, Philip (2016): Articulatory and
  Acoustic Characteristics of German Fricative Clusters, Phonetica, 73(1),
  52--78.

Cederbaum, Pouplier, Hoole, Greven (2016): Functional Linear Mixed
  Models for Irregularly or Sparsely Sampled Data. Statistical Modelling,
  16(1), 67-88.

Jona Cederbaum (2019). sparseFLMM: Functional Linear Mixed Models for
  Irregularly or Sparsely Sampled Data. R package version 0.3.0.
   \url{https://CRAN.R-project.org/package=sparseFLMM}
}
\usage{
phonetic_subset
}
\description{
A small subset of the phonetics data set \code{\link[multifamm]{phonetic}}
with observations from two speakers and two items only. This will not produce
meaningful results but can be used as a toy data set when testing the code.
The variables are as in the full data set, see
\code{\link[multifamm]{phonetic}}
}
\keyword{datasets}
