% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_simulation.R
\name{predict_funData}
\alias{predict_funData}
\title{Predict funData Object for Unobserved Values}
\usage{
predict_funData(funData, new)
}
\arguments{
\item{funData}{funData object to be evaluated.}

\item{new}{Vector containing the time points to be approximated.}
}
\description{
This is an internal function. It takes a funData object and approximates the
function values for the supplied new time points using the adapted approxNA()
function. The function checks whether the time points in the argument new are
already included in the argvals. Predict is probably not the right word but
it is used for consistency.
}
