% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_extract.R
\name{predict_mean}
\alias{predict_mean}
\title{Predict The Mean Function For the FPC Plots}
\usage{
predict_mean(model, multi = TRUE, dimnames = c("aco", "epg"))
}
\arguments{
\item{model}{multiFAMM model or list of univariate models for which to
predict the mean.}

\item{multi}{Indicator if it is a multiFAMM model (TRUE) or a list of
univariate models.}

\item{dimnames}{Vector of strings containing the names of the dimensions.}
}
\value{
A multiFunData object.
}
\description{
This is an internal function that helps to interpret the FPCs. Extract the
mean function for all covariates set to 0.5. This is useful if combined with
the estimated FPCs because one can then add and subtract suitable multiples
from this function.
}
