% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_results.R
\name{predict_sparseFLMM_covar}
\alias{predict_sparseFLMM_covar}
\title{Predict The Covariate Effects of a Univariate sparseFLMM Model}
\usage{
predict_sparseFLMM_covar(model, type = "terms", unconditional = FALSE,
  grid = seq(0, 1, length.out = 100))
}
\arguments{
\item{model}{sparseFLMM model with covariates.}

\item{type}{Gam terms to be predicted.}

\item{unconditional}{Std conditional on lambda.}

\item{grid}{Grid of evaluation points. Defaults to observations on [0,1].}
}
\description{
This is an internal function that helps to compare univariate sparseFLMM
models to multiFAMM models. This functions predicts the covariate effects of
a model that has the structure of a regression model fitted by sparseFLMM().
}
\details{
This functions' name in the thesis was predict_model().
}
