% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_results.R
\name{predictedUnivar2DataFrame}
\alias{predictedUnivar2DataFrame}
\title{Transform Output of Covariate Predictions of Univariate Models to Data Frame}
\usage{
predictedUnivar2DataFrame(
  aco_pr,
  epg_pr,
  effect,
  grid = seq(0, 1, length.out = 100)
)
}
\arguments{
\item{aco_pr}{Output of the first model (dimension aco).}

\item{epg_pr}{Output of the second model (dimension epg). Can be NULL.}

\item{effect}{Which effect to extract. If intercept is specified (1), then
the scalar intercept is added.}

\item{grid}{Grid of evaluation points. Defaults to observations on [0,1].}
}
\description{
This is an internal function that helps to compare univariate sparseFLMM
models to multiFAMM models. This functions converts the predictions of the
covariate effects of one or two models as given by the function
predict_sparseFLMM_covar() to a data.frame.
}
\details{
This functions' name in the thesis was predict2DataFrame().
}
