% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_simulation.R
\name{prepare_gam_predict}
\alias{prepare_gam_predict}
\title{Prepare Prediction of Simulated Data Using MGCV Model Object}
\usage{
prepare_gam_predict(data, num_cov, interaction, which_inter, model)
}
\arguments{
\item{data}{Simulated data set for which the necessary model components are
to be computed.}

\item{num_cov}{Number of covariates if covariate = TRUE.}

\item{interaction}{TRUE if there are interactions between covariates (as in
sparseFLMM).}

\item{which_inter}{Symmetric matrix specifying the interaction terms (as in
sparseFLMM).}

\item{model}{GAM model from which to extract the covariate functions.}
}
\description{
This is an internal function. It is used to generate new data sets. It adapts
curve_info so that it has the same structure as the original data. Then, it
is possible to use the predict function from mgcv.
}
