% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfpca.R
\name{prepare_mfpca}
\alias{prepare_mfpca}
\title{Prepare Information Necessary for MFPCA}
\usage{
prepare_mfpca(model_list, fRI_B, mfpc_weight)
}
\arguments{
\item{model_list}{List containing sparseFLMM objects for each dimension as
given by the output of apply_sparseFLMM()}

\item{fRI_B}{Boolean for including functional random intercept for individual
(B in Cederbaum). Defaults to \code{FALSE}.}

\item{mfpc_weight}{TRUE if the estimated univariate error variance is to be
used as weights in the scalar product of the MFPCA.}
}
\description{
This is an internal function contained in the multiFAMM function. This step
uses the information from the univariate FLMMs for the MFPCA. It also allows
a simple weighting scheme of the MFPCA.
}
\keyword{internal}
