% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_famm.R
\name{refit_for_weights}
\alias{refit_for_weights}
\title{Refit the model under an independence assumption}
\usage{
refit_for_weights(formula, data, model_list)
}
\arguments{
\item{formula}{Formula to fit the final model.}

\item{data}{Data that contains all the variables specified in formula.}

\item{model_list}{List containing sparseFLMM objects for each dimension}
}
\description{
This is an internal function. Refit the model under an independence
assumption now with the basis functions from the MFPCA. Goal is to extract an
estimate for the error variances.
}
