% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_results.R
\name{return_number_fpcs}
\alias{return_number_fpcs}
\title{Return Number of Selected FPCS per Component and Iteration}
\usage{
return_number_fpcs(folder, uni = FALSE)
}
\arguments{
\item{folder}{Folder with saved objects from the simulation.}

\item{uni}{Extract also the univariate part of the components. Defaults to
FALSE.}
}
\description{
This is an internal function. It takes the a folder containing simulation
results and returns a data set which contains the number of selected FPCS
per component.
}
