% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_results.R
\name{rrMSE}
\alias{rrMSE}
\title{Root (Relative) Mean Squared Error for Scalar Estimates}
\usage{
rrMSE(theta_true, theta_estim, relative = TRUE)
}
\arguments{
\item{theta_true}{True component (scalar).}

\item{theta_estim}{Estimated component (scalar).}

\item{relative}{FALSE if the absolute MSE is to be computed. Defaults to the
relative MSE (TRUE).}
}
\description{
This is an internal function. It calculates the root (relative) mean squared
error for scalar quantities.
}
