% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_results.R
\name{sim_eval_components}
\alias{sim_eval_components}
\title{Evaluate the model components in the simulation}
\usage{
sim_eval_components(folder, m_true_comp, label_cov, relative = TRUE, I = 9,
  J = 16, reps = 5, nested = FALSE, fixed_fpc = TRUE, sigma2 = NULL)
}
\arguments{
\item{folder}{Folder with saved objects from the simulation.}

\item{m_true_comp}{True model components used for the simulation.}

\item{label_cov}{Vector of labels for the covariates.}

\item{relative}{Is the relative error measure to be used or the absolute?
Defaults to TRUE.}

\item{I}{Number of levels for component B. Defaults to 9.}

\item{J}{Number of levels for component C. Defaults to 16.}

\item{reps}{Number of repetitions of the levels of B and C. Defaults to 5.}

\item{nested}{TRUE if the model component C is nested. Defaults to FALSE.}

\item{fixed_fpc}{FALSE if the number of FPCs in the model is selected by the
model not by the user. This means that eigenvalues, scores and
eigenfunctions are not evaluated. Defaults to TRUE.}

\item{sigma2}{Vector of true values can be supplied if they differ from the
values in m_true_comp. Defaults to NULL.}
}
\description{
This is an internal function. The function takes the folder containing the
results of a simulation scenario and returns a data frame with relative
MSE values for the separate model components.
}
