% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_results.R
\name{sim_eval_dimensions}
\alias{sim_eval_dimensions}
\title{Evaluate the model components per dimension in the simulation}
\usage{
sim_eval_dimensions(folder, m_true_comp, label_cov, relative = TRUE,
  uni_compare = FALSE, I = 9, J = 16, reps = 5, nested = FALSE)
}
\arguments{
\item{folder}{Folder with saved objects from the simulation.}

\item{m_true_comp}{True model components used for the simulation.}

\item{label_cov}{Vector of labels for the covariates.}

\item{relative}{Is the relative error measure to be used or the absolute?
Defaults to TRUE.}

\item{uni_compare}{TRUE if the simulation scenario includes univariate model
estimates that should be also evaluated. Defaults to FALSE.}

\item{I}{Number of levels for component B. Defaults to 9.}

\item{J}{Number of levels for component C. Defaults to 16.}

\item{reps}{Number of repetitions of the levels of B and C. Defaults to 5.}

\item{nested}{TRUE if the model component C is nested. Defaults to FALSE.}
}
\description{
This is an internal function. The function takes the folder containing the
results of a simulation scenario and returns a data frame with relative
MSE values for the separate model components and the separate dimensions.
}
