% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_results.R
\name{urrMSE}
\alias{urrMSE}
\title{Univariate Root (Relative) Mean Squared Error for Functions}
\usage{
urrMSE(fun_true, fun_estim, flip = FALSE, relative = TRUE)
}
\arguments{
\item{fun_true}{MultiFunData object containing the true functions.}

\item{fun_estim}{MultiFunData object containing the estimated functions.}

\item{flip}{Are the estimated functions to be flipped before calculating the
mrrMSE? Defaults to FALSE.}

\item{relative}{FALSE if the absolute MSE is to be computed. Defaults to the
relative MSE (TRUE).}
}
\description{
This is an internal function. Calculate the univariate root (relative) mean
squared error for functions. For eigenfunctions, which are only defined up to
a sign change, there is the option to flip the functions first if the flipped
functions seems to be closer to the true function
}
