% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_mfpcs.R
\name{compute_var}
\alias{compute_var}
\title{Compute the Number of FPCs needed}
\usage{
compute_var(sigma_sq, values, mfpc_cutoff, tot_var)
}
\arguments{
\item{sigma_sq}{Vector containing the estimated variances on each dimension.}

\item{values}{List containing the multivariate Eigenvalues for each variance
component.}

\item{mfpc_cutoff}{Pre-specified level of explained variance of results of
MFPCA.}

\item{tot_var}{Scalar with the value of the total variance as computed by the
univariate sparseFLMM.}
}
\description{
This is an internal function. The function takes all the information needed
to calculate how many FPCs are needed to reach the pre-specified cutoff
level.
}
