% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_plots.R
\name{covariate_plot_helper}
\alias{covariate_plot_helper}
\title{Extract Data for Covariate Plot}
\usage{
covariate_plot_helper(
  model,
  mcomp,
  dimlabels,
  int_include = TRUE,
  two_d = FALSE,
  multi = TRUE,
  m_fac = 1.96
)
}
\arguments{
\item{model}{The multifamm model from which to extract the estimates.}

\item{mcomp}{The extracted model components as given by the internal function
extract_components().}

\item{dimlabels}{The dimensions of the model.}

\item{int_include}{Include the estimate and uncertainty of the scalar
intercept in the values of the functional intercept. Defaults to TRUE.}

\item{two_d}{Is a two dimensional representation (e.g. x and y coordinate) a
better representation for the data. Defaults to FALSE (one variable).}

\item{multi}{Is the model a multifamm or a sparseFLMM model. Defaults to TRUE
(multifamm model).}

\item{m_fac}{Multiplication factor to represent the confidence interval of
the estimates.}
}
\description{
This is an internal function. It gives as an output the data set necessary to
plot the covariate effects. It is meant to be general enough to also include
more than two dimensions.
}
