% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_results.R
\name{load_sim_results}
\alias{load_sim_results}
\title{Load All Simulation Results of Model Components of a Folder}
\usage{
load_sim_results(folder, component)
}
\arguments{
\item{folder}{Folder with saved objects from the simulation.}

\item{component}{String of the model components to be loaded.}
}
\description{
This is an internal function. It takes the a folder containing simulation
results and what part of the model is to be loaded. It then loads all
corresponding results and concatenates them.
}
