% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfunctions_results.R
\name{sim_eval_dimensions}
\alias{sim_eval_dimensions}
\title{Evaluate the model components per dimension in the simulation}
\usage{
sim_eval_dimensions(folder, m_true_comp, label_cov, uni_compare = FALSE)
}
\arguments{
\item{folder}{Folder with saved objects from the simulation.}

\item{m_true_comp}{True model components used for the simulation.}

\item{label_cov}{Vector of labels for the covariates.}

\item{uni_compare}{TRUE if the simulation scenario includes univariate model
estimates that should be also evaluated. Defaults to FALSE.}
}
\description{
This is an internal function. The function takes the folder containing the
results of a simulation scenario and returns a data frame with relative
MSE values for the separate model components and the separate dimensions.
}
